Introduce `ack-generate webhooks` subcommand
This patch brings in a new sub command that can be used to generate
webhook functions. For now it can only generate conversion webhook
functions, but it can be used to generate defaulting and validating
functions.

This patch also introduces a new function to load existing ack-generate
metadata for a given api version.